backend/pyproject.toml
[project]
name = "nimbus"
version = "0.1.0"
description = "Nimbus - Pulumi-based SaaS Control Panel"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic[email]>=2.10.0",
    "pydantic-settings>=2.7.0",
    "pyyaml>=6.0.2",
    "python-jose[cryptography]>=3.3.0",
    "argon2-cffi>=23.1.0",
    "strawberry-graphql[fastapi]>=0.254.0",
    "temporalio>=1.9.0",
    "httpx>=0.28.0",
    "minio>=7.2.0",
    "authlib>=1.3.0",
    "jinja2>=3.1.0",
    "aiosmtplib>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "httpx>=0.28.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "RUF"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
